feat: team driven player spawn

+++ Assets/BeMyShotgunSir/Scripts/Core/Lobby/LobbyManager.cs
@@ -26,7 +26,7 @@ namespace BeMyShotgunSir.Scripts.Core.Lobby
         void SetLobbyInfo_Response(LobbyInfo info);
         void SetPlayerCount_Response(int prev, int next);
         void SetPlayerStates_Response(SyncDictionaryOperation op, int key, LobbyPlayerState value);
-        void SetTeamInfos_Response(SyncDictionaryOperation op, long key, LobbyTeamInfo value);
+        void SetTeamInfos_Response(SyncDictionaryOperation op, int key, LobbyTeamInfo value);
     }
 
     public interface ILobbyManager : IManager, ILobbyManager_Bootstrapper, ILobbyManager_NetController { }
@@ -193,6 +193,6 @@ namespace BeMyShotgunSir.Scripts.Core.Lobby
         }
         public void SetPlayerStates_Response(SyncDictionaryOperation op, int key, LobbyPlayerState value) => _viewModel.SetPlayerStates(op, key, value);
 
-        public void SetTeamInfos_Response(SyncDictionaryOperation op, long key, LobbyTeamInfo value) => _viewModel.SetTeamInfos(op, key, value);
+        public void SetTeamInfos_Response(SyncDictionaryOperation op, int key, LobbyTeamInfo value) => _viewModel.SetTeamInfos(op, key, value);
     }
 }
